third_party/WebKit/LayoutTests/fast/multicol/span/offset-properties.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="position:relative; border:10px solid; columns:3; column-gap:0; column-fill:auto; width:300px;">
    <div style="height:450px;"></div>
    <div id="elm" style="column-span:all; margin:1px; border:10px solid; height:50px;"></div>
</div>
<script>
    test(function() {
        assert_equals(document.getElementById('elm').offsetLeft, 1);
        assert_equals(document.getElementById('elm').offsetTop, 151);
    }, "offsetLeft and offsetTop on column-span:all");
</script>
